Separated blocks and bricks into separate classes
